pyproject.toml
[project]
name = "interactive_sketchpad"
version = "0.0.1"
description = "Interactive Sketchpad"
readme = "README.md"
authors = [
    { name = "Steven-Shine Chen", email = "stevensh@mit.edu" }
]
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.0.3",
    "openai>=1.55.0",
    "pillow>=11.0.0",
    "pydantic>=2.10.1",
    "python-dotenv>=1.0.1",
    "scikit-image>=0.25.0",
    "tenacity>=9.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black>=24.10.0",
    "datasets>=3.1.0",
    "ipykernel>=6.29.5",
    "isort>=5.13.2",
    "mlflow>=2.19.0",
    "pre-commit>=4.0.1",
    "pytest>=8.3.4",
    "tiktoken>=0.8.0",
]
